# openapi.yaml
openapi: 3.0.3
info:
  title: Kanban System API
  description: 看板工作流管理系统 API
  version: 2.0.0
servers:
  - url: /api
    description: API Server
...
paths:
...
  /cards/{id}/checklist:
    get:
      tags:
        - checklist
      summary: 获取卡片的检查清单
      operationId: getChecklist
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChecklistItem'
    post:
      tags:
        - checklist
      summary: 创建检查清单项
      operationId: createChecklistItem
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChecklistItemRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChecklistItem'
  /checklist/{id}:
    put:
      tags:
        - checklist
      summary: 更新检查清单项
      operationId: updateChecklistItem
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChecklistItemRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChecklistItem'
    delete:
      tags:
        - checklist
      summary: 删除检查清单项
      operationId: deleteChecklistItem
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # 评论接口
  /cards/{id}/comments:
    get:
      tags:
        - comment
      summary: 获取卡片的评论
      operationId: getComments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
    post:
      tags:
        - comment
      summary: 创建评论
      operationId: createComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
  /comments/{id}:
    put:
      tags:
        - comment
      summary: 更新评论
      operationId: updateComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
    delete:
      tags:
        - comment
      summary: 删除评论
      operationId: deleteComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # 标签接口
  /boards/{id}/labels:
    get:
      tags:
        - label
      summary: 获取看板的标签
      operationId: getLabels
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
    post:
      tags:
        - label
      summary: 创建标签
      operationId: createLabel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLabelRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
  /labels/{id}:
    put:
      tags:
        - label
      summary: 更新标签
      operationId: updateLabel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLabelRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
    delete:
      tags:
        - label
      summary: 删除标签
      operationId: deleteLabel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # 附件接口
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        nickname:
          type: string
        avatar:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
...
    ChecklistItem:
      type: object
      properties:
        id:
          type: integer
        card_id:
          type: integer
        content:
          type: string
        completed:
          type: boolean
        position:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - card_id
        - content
        - completed
        - position
    Label:
      type: object
      properties:
        id:
          type: integer
        board_id:
          type: integer
        name:
          type: string
        color:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - board_id
        - name
        - color
    Comment:
      type: object
      properties:
        id:
          type: integer
        card_id:
          type: integer
        user_id:
          type: integer
        user:
          $ref: '#/components/schemas/User'
        content:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - card_id
        - user_id
        - content
...
    CreateChecklistItemRequest:
      type: object
      properties:
        content:
          type: string
          maxLength: 200
      required:
        - content
    UpdateChecklistItemRequest:
      type: object
      properties:
        content:
          type: string
          maxLength: 200
        completed:
          type: boolean
        position:
          type: integer
    CreateCommentRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    UpdateCommentRequest:
      type: object
      properties:
        content:
          type: string
    CreateLabelRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
        color:
          type: string
          maxLength: 7
      required:
        - name
        - color
    UpdateLabelRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
        color:
          type: string
          maxLength: 7
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message